+++ src/scenes/settings_scene.lua
@@ -1,4 +1,5 @@
-local colors = require "src.consts.colors"
+local inputManager  = require("src.managers.input_manager")
+local colors        = require("src.consts.colors")
 
 local settingsScene = {
     assets = {},
@@ -12,14 +13,14 @@ function settingsScene:load(assets, actions, configs)
     self.configs = configs
 end
 
-function settingsScene:keypressed(key)
-    if key == "escape" then
+function settingsScene:handleInputs()
+    if inputManager:wasPressed("back") then
         self.actions.switchScene("title")
     end
 end
 
 function settingsScene:update(dt)
-
+    self:handleInputs()
 end
 
 function settingsScene:draw()